(#33) Реализовал логику для отображения данных о полученном городе
