+++ src/components/weather-card/weather-card.css
@@ -5,33 +5,3 @@
   padding: 15px;
   box-shadow: 0px 2px 11px 0px #000;
 }
-
-.weather-card__content_display {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  max-width: 250px;
-  max-height: 90px;
-  margin: 40px;
-  margin-top: 60px;
-}
-
-.weather-card__content_input {
-  width: 100%;
-  border: none;
-  font-size: 16px;
-  padding: 5px;
-}
-
-.weather-card__content_emoji_p {
-  font-size: 90px;
-  margin-right: 50px;
-}
-
-.weather-card__content_temp_p {
-  font-size: 70px;
-}
-
-.capitalize {
-  text-transform: capitalize;
-}
